Cast needed

+++ gecode/kernel/memory/region.hpp
@@ -61,7 +61,8 @@ namespace Gecode {
       /// Amount of free memory
       size_t free;
       /// The actual memory area (allocated from top to bottom)
-      alignas(std::max(alignof(std::max_align_t),GECODE_MEMORY_ALIGNMENT))
+      alignas(std::max(alignof(std::max_align_t),
+                       static_cast<size_t>(GECODE_MEMORY_ALIGNMENT)))
         double area[Kernel::MemoryConfig::region_area_size / sizeof(double)];
       /// A pointer to another chunk
       Chunk* next;
@@ -333,7 +334,8 @@ namespace Gecode {
   forceinline bool
   Region::Chunk::alloc(size_t s, void*& p) {
     Kernel::MemoryConfig::align
-      (s,std::max(alignof(std::max_align_t),GECODE_MEMORY_ALIGNMENT));
+      (s,std::max(alignof(std::max_align_t),
+                  static_cast<size_t>(GECODE_MEMORY_ALIGNMENT)));
     if (s > free)
       return false;
     free -= s;
